Initialize project structure under releases/release-1.3/v1.3.4
