--- library.properties
@@ -1,10 +1,0 @@
-name=Firmus
-version=1.3.3
-author=Keshav Haripersad
-maintainer=kharipersad11.h@gmail.com
-sentence=todo
-paragraph=todo
-category=Motion
-url=https://github.com/Keshav11-coder/Firmus-SDK
-architectures=esp32
-includes=Firmus, Firmus/Assets
