+++ releases/release-1.3/v1.3.4/library.properties
@@ -1,0 +1,10 @@
+name=Firmus
+version=1.3.3
+author=Keshav Haripersad
+maintainer=kharipersad11.h@gmail.com
+sentence=todo
+paragraph=todo
+category=Motion
+url=https://github.com/Keshav11-coder/Firmus-SDK
+architectures=esp32
+includes=Firmus, Firmus/Assets
