+++ releases/release-1.3/v1.3.4/tests/QuadCascadeCtrlAtt_RM4_v1/QuadCascadeCtrlAtt_RM4_v1.ino
@@ -1,0 +1,127 @@
+// © 2025 Keshav Haripersad
+//
+// Footprint:
+// - Heap Usage: ~35.4 KB used / 303 KB total (~11.7% utilization)
+// - CPU Usage: ~6.5% (single-core approximation, 1 Hz loop rate)
+// - MCU: ESP32 (dual-core, 240 MHz)
+// - Sensors: MPU6050 (I2C interface)
+// - Control: Cascade PID attitude controller with RigidModel<4>
+// - Loop Frequency: ~1 Hz (timing based on millis())
+// - Features: Real-time orientation & rate feedback, torque computation
+// - Notes: Efficient static allocation, minimal dynamic memory usage
+// - Limitations: CPU usage measured only on Core 0; does not include multi-core workload
+// - Usage: Initialize MPU6050, calibrate sensor, call controller compute(dt) each loop
+// - Logging: Serial prints CPU usage every second; disable or adapt for higher frequencies
+//
+// References:
+// - SDK v1.3.1
+// - ESP-IDF timers and FreeRTOS primitives for CPU measurement
+// - MPU6050 datasheet and driver documentation
+
+#include <dronev2.h>
+#include <dronev2/mpu6050.h>
+#include <dronev2/Controllers/Cascade.h> // Uses RigidModel by default
+
+using namespace dronev2;
+using namespace Matrix2;
+
+const float theta_1 = PI / 4;
+const float theta_2 = (3 * PI) / 4;
+const float theta_3 = (5 * PI) / 4;
+const float theta_4 = (7 * PI) / 4;
+
+const float l = 0.125;
+
+const float k_yaw = 1;
+const float k_F = 1;
+
+auto model4 = RigidModel<4>()
+  .setMomentArm(l, l, l, l)
+  .setActuatorAngles(theta_1, theta_2, theta_3, theta_4)
+  .setAllocationMatrix({
+    { l * sin(theta_1), l * sin(theta_2), l * sin(theta_3), l * sin(theta_4) }, // torque x
+    { l * cos(theta_1), l * cos(theta_2), l * cos(theta_3), l * cos(theta_4) }, // torque y
+    { -k_yaw,  k_yaw,  -k_yaw,   k_yaw }, // torque z
+    {  k_F,   k_F,    k_F,    k_F  } // thrust T
+  })
+  .setInertiaMatrix(diagonal<float, 3>({ 0.125f, 0.125f, 0.10f }))
+  .setMass(0.8)
+  .setOperatingVoltage(11.1)
+  .setMotorKv(2300, 2300, 2300, 2300)
+  .setPropellerDiameter(0.1016)
+  .setPropellerPitch(0.1270);
+
+auto attitudeCtrl = Cascade::Attitude()
+  .configureOrientationGains(PIDGains{.kp = 4.5, .ki = 0.02, .kd = 0.05, .kf = 15},
+                             PIDGains{.kp = 4.5, .ki = 0.02, .kd = 0.05, .kf = 15},
+                             PIDGains{.kp = 5.0, .ki = 0.10, .kd = 0.05, .kf = 10})
+  .configureRateGains(PIDGains{.kp = 2.5, .ki = 0.05, .kd = 0.02, .kf = 10},
+                      PIDGains{.kp = 2.5, .ki = 0.05, .kd = 0.02, .kf = 10},
+                      PIDGains{.kp = 3.5, .ki = 0.02, .kd = 0.01, .kf = 10});
+
+auto controller = Cascade::Controller()
+  .useModel(&model4)
+  .useAttitudeInterface(&attitudeCtrl);
+
+mpu6050 mpu(19, 23);
+
+void setup() {
+  Serial.begin(115200);
+
+  mpu.begin();
+  mpu.calibrate();
+
+  Cascade::Attitude::ControlFlags flags;
+  flags.useTiltLimit = true;
+  flags.useRateLimit = true;
+  flags.holdHeading = false;
+
+  flags.tiltLimitAbs = PI / 4;
+  flags.rateLimitAbs = 3.0f;
+
+  attitudeCtrl.setControlFlags(&flags);
+}
+
+unsigned long last_time = 0;
+float dt = 0;
+
+void loop()
+{
+  unsigned long now = micros();
+  dt = (now - last_time) / 1000000.0f;
+  last_time = now;
+
+  mpu.read();
+
+  attitudeCtrl.updateMeasuredOrientation(mpu.pitch_rad, mpu.roll_rad, mpu.yaw_rad);
+  attitudeCtrl.updateMeasuredRate(mpu.gx, mpu.gy, mpu.gz);
+
+  Cascade::Controller::Torque torque = controller.compute(dt);
+
+  Serial.print("Torque X: ");
+  Serial.print(torque.x);
+  Serial.print(" Torque Y: ");
+  Serial.print(torque.y);
+  Serial.print(" Torque Z: ");
+  Serial.println(torque.z);
+}
+
+// Sketch Promotion Checklist:
+// Main focus: Abstracting the SDK with Attitude / Controller as first focus.
+//
+// [X] Verify attitude controller PID/PID_filtered gains correctly configured and stable.
+// [X] Confirm all relevant control classes are derived from base Interface classes and comply with architecture design.
+//     - IRigidModel, IAttitude, IVelocity, IPosition, IController must be architecturally complete
+//     - No hardcoding
+// [X] Confirm official Cascade::Attitude and Cascade::Controller implementations with no issues.
+// [X] Ensure timing loop accurately computes dt (delta time) in seconds with no overflow issues.
+// [X] Verify heap usage remains stable under typical and peak loads (no leaks) ~ < 40kB.
+// [X] Confirm torque computations produce expected ranges and no NaNs/infs occur.
+// [X] Stress test loop at target frequencies (1 Hz and higher) without missed deadlines or crashes.
+// [X] Document any deviations from expected behavior and track fixes.
+// [X] Code fully compiles with no warnings or errors on new SDK version.
+// [X] Confirm all third-party dependencies (dronev2, MPU drivers) updated and compatible.
+// [X] Review and update all relevant documentation and comments for new SDK changes.
+//
+// [X] Final sign-off by engineering lead before SDK promotion.
+//     SDK 1.3.1 -> 1.3.2
