--- src/Firmus.cpp
@@ -1,18 +1,0 @@
-/** © 2025 Keshav Haripersad
- *  Base API logic - Firmus.cpp
- *  | Check out Firmus.h for an overview.
- *  | Check out files in /Firmus for other modules.
- *
- *  Licensed under the Apache 2.0 license (check LICENSE).
- *  Github link: https://github.com/Keshav11-coder/Firmus-SDK
- *  Firmus SDK version: 1.3.3
- */
-
-// Headers
-#include "Firmus.h"
-
-// Currently unused, but can be used in the future.
-namespace Firmus
-{
-    
-}
